+++ src/djangoflash/tests/suite.py
@@ -5,7 +5,6 @@
 
 import sys
 
-
 # Adds the Django test project to system path
 from django.core.management import setup_environ
 import djangoflash.tests.testproj.settings as project_settings
@@ -18,7 +17,9 @@ setup_test_environment()
 connection.creation.create_test_db()
 
 # Imports unit tests
+from context_processors import *
 from models import *
+from storage import *
 
 # Imports integration tests
 from testproj.app.tests import *