+++ src/djangoflash/tests/testproj/app/tests.py
@@ -7,7 +7,7 @@ from django.core.urlresolvers import reverse
 from django.test import TestCase
 
 from djangoflash.context_processors import CONTEXT_VAR
-from djangoflash.middleware import FlashMiddleware
+from djangoflash.middleware import FlashScope
 
 from testproj.app import views
 
@@ -20,23 +20,6 @@ class IntegrationTestCase(TestCase):
         """Shortcut to get the flash scope from the view context.
         """
         return self.response.context[CONTEXT_VAR]
-
-    def test_session_state_for_unused_flash(self):
-        """Flash scope shouldn't be stored in the session if there's no flash.
-        """
-        self.response = self.client.get(reverse(views.render_template))
-        self.assertFalse(FlashMiddleware._SESSION_KEY in self.client.session)
-
-    def test_session_state_for_used_flash(self):
-        """Flash scope should be removed from the session if there's no flash.
-        """
-        self.response = self.client.get(reverse(views.set_flash_var))
-        self.response = self.client.get(reverse(views.render_template))
-        self.assertTrue(FlashMiddleware._SESSION_KEY in self.client.session)
-
-        # Flash scope should be removed from the session
-        self.response = self.client.get(reverse(views.render_template))
-        self.assertFalse(FlashMiddleware._SESSION_KEY in self.client.session)
 
     def test_default_lifecycle(self):
         """A value should be automatically removed from the flash scope.
@@ -115,7 +98,7 @@ class IntegrationTestCase(TestCase):
         self.assertRaises(TypeError, request)
 
     def test_remove_flash_scope(self):
-        """The app should not break even if request.flash doesn't exists.
+        """The application should handle inexistent flash scope properly.
         """
         self.response = self.client.get(reverse(views.remove_flash))
-        self.assertFalse(self._flash() is None)
+        self.assert_(self._flash() != None)